fix(test): move thread-heavy tests from Tier 1 to Tier 2 integration
Thread-spawning and timing-dependent tests in tests/unit/runtime/ were
exhausting CI runner thread limits after ~5000 tests, causing cascading
timeouts across all Python versions. These tests violated the Tier 1
contract (fast, isolated, deterministic).

Moved 9 whole files and extracted thread-heavy test classes/methods from
4 partial files into tests/tier2_integration/runtime/. Removed all
pytestmark slow/performance markers from moved files. Added a dedicated
Tier 2 CI step with appropriate 30s timeout. Removed continue-on-error
for Python 3.13 so CI is strict on all 3 versions.

Local verification: 5605 unit tests pass in 37s, 258 tier2 tests pass
in 5s, zero failures.
