+++ .github/workflows/unified-ci.yml
@@ -31,9 +31,6 @@ concurrency:
   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.run_id }}
   cancel-in-progress: true
 
-env:
-  SKIP_SLOW_TESTS: true
-
 jobs:
   # Skip push CI when an open PR already covers the branch
   check-duplicate:
@@ -70,7 +67,7 @@ jobs:
     if: needs.check-duplicate.outputs.should-skip != 'true'
     runs-on: ubuntu-latest
     timeout-minutes: 15
-    continue-on-error: ${{ matrix.python-version == '3.13' }}
+    continue-on-error: false
     strategy:
       fail-fast: false
       matrix:
@@ -102,14 +99,22 @@ jobs:
           .venv/bin/pyright src/kailash/ --pythonversion ${{ matrix.python-version }} --level error 2>&1 | tail -5
 
       - name: Test (Tier 1 — unit tests)
-        timeout-minutes: 15
+        timeout-minutes: 10
         run: |
           .venv/bin/python -m pytest \
             tests/unit/ \
             tests/trust/plane/unit/ \
             tests/security/ \
-            -m 'not (slow or integration or e2e or performance or requires_docker or requires_postgres or requires_mysql or requires_redis or requires_ollama)' \
-            --maxfail=10 -q -p no:xdist --timeout=30
+            -m 'not (slow or performance or integration or e2e or requires_docker or requires_postgres or requires_mysql or requires_redis or requires_ollama)' \
+            --maxfail=10 -q -p no:xdist --timeout=10
+
+      - name: Test (Tier 2 — integration runtime tests)
+        timeout-minutes: 10
+        run: |
+          .venv/bin/python -m pytest \
+            tests/tier2_integration/ \
+            -m 'not (requires_docker or requires_postgres or requires_mysql or requires_redis or requires_ollama)' \
+            --maxfail=5 -q -p no:xdist --timeout=30
 
   test-pact:
     name: Test PACT (Python 3.11)
