+++ tests/tier2_integration/runtime/test_distributed_runtime.py
@@ -13,8 +13,6 @@ from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 import pytest
 
-# Distributed runtime tests exhaust CI runner threads
-pytestmark = pytest.mark.slow
 
 from src.kailash.runtime.distributed import (
     DistributedRuntime,
