+++ tests/tier2_integration/runtime/test_double_execution_debug.py
@@ -10,8 +10,6 @@ import time
 
 import pytest
 
-# Cyclic workflow tests exhaust CI runner threads
-pytestmark = pytest.mark.slow
 
 from kailash.runtime.local import LocalRuntime
 from kailash.workflow.builder import WorkflowBuilder
