+++ tests/tier2_integration/runtime/test_malformed_output_validation.py
@@ -8,8 +8,6 @@ from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 
-# Cycle tests exhaust CI runner threads
-pytestmark = pytest.mark.slow
 
 from kailash.runtime.local import LocalRuntime
 from kailash.workflow.builder import WorkflowBuilder
@@ -19,7 +17,6 @@ from kailash.workflow.cyclic_runner import CyclicWorkflowExecutor, WorkflowState
 class TestMalformedOutputValidation:
     """Test cases for validating and handling malformed node outputs in cyclic workflows."""
 
-    @pytest.mark.slow
     def test_none_exit_result_in_cycle_termination(self):
         """Test handling when exit_result is None in cycle termination logic."""
         # This specifically tests the bug on line 882 of cyclic_runner.py
