+++ tests/tier2_integration/runtime/test_phase0d_optimizations.py
@@ -11,8 +11,6 @@ import time
 
 import pytest
 
-# Performance tests with tight timing assertions fail on CI runners
-pytestmark = pytest.mark.performance
 
 from kailash.runtime.local import LocalRuntime
 from kailash.tracking.metrics_collector import MetricsCollector, MetricsContext
