+++ tests/tier2_integration/runtime/test_phase0e_optimizations.py
@@ -16,8 +16,7 @@ import sqlite3
 
 import pytest
 
-# SQLite I/O tests fail on CI runners with disk I/O errors
-pytestmark = pytest.mark.performance
+
 import tempfile
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
