+++ tests/tier2_integration/runtime/test_resource_limit_enforcer.py
@@ -22,8 +22,6 @@ from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 
-# Thread-heavy tests exhaust CI runner limits
-pytestmark = pytest.mark.slow
 
 # Import will be added once we create the class
 # from kailash.runtime.resource_manager import ResourceLimitEnforcer, EnforcementPolicy
