+++ tests/tier2_integration/runtime/test_retry_policy_engine.py
@@ -20,8 +20,6 @@ from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
 
-# Thread-heavy tests exhaust CI runner limits
-pytestmark = pytest.mark.slow
 
 from kailash.runtime.resource_manager import (
     AdaptiveRetryStrategy,
